language now uses i32 and f32 instead of i64 and f64

+++ src/common/mod.rs
@@ -3,7 +3,7 @@ pub mod op;
 
 #[derive(Debug, Clone, Copy, PartialEq)]
 pub enum Literal {
-    Int(i64),
-    Float(f64),
+    Int(i32),
+    Float(f32),
     Bool(bool)
 }
